+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### 功能构建
 
+- 增加预设查询。
+  - com.dwarfeng.judge.stack.service.SinkerInfoMaintainService.KEY_NOT_IN。
+
 - 优化部分类的日志输出。
   - com.dwarfeng.judge.impl.handler.AnalysisOperateHandlerImpl。